framework/tabular/requirements.txt

# for modelling
lightgbm
fairlearn
xlrd
seaborn
sklearn-compat==0.1.3
scikit-learn==1.6.1
lightgbm==3.3.0
#imbalanced-learn==0.13.0
gdown
jax
optax
flax
ipywidgets

# for fairness
holisticai==1.0.14

# for explainability
pyyaml
interpret
aix360
lime

# for AI documentation
model-card-toolkit

# for deployment
kserve==0.15.2
mlflow==3.2.0   #  mlflow[extras]  To deploy with MLServer
mlserver==1.7.1
mlserver_mlflow==1.7.1